keep executor in jobs.go
